- moved workspace sidebar card to it's own component  - setting dummy 'last_used_workspace_id' to `Uuid::nil()` when the app is self-hosted  - added `AppType` struct to determine if the app is self-hosted or managed  - conditionally rendering parts of login, sign up, create deployment, and deployment list pages depending on the `AppType`
