+++ hosted-frontend/src/queries/infrastructure/deployment/mod.rs
@@ -12,7 +12,7 @@ pub fn list_deployments_query(
 ) -> QueryScope<AllDeploymentsTag, Result<ListDeploymentResponse, ServerFnError<ErrorType>>> {
 	let (state, _) = AuthState::load();
 	let access_token = state.get().get_access_token();
-	// TODO remove this unwrap
+	// TODO: remove this unwrap
 	let workspace_id = state.get().get_last_used_workspace_id().unwrap();
 
 	create_query(
