+++ hosted-frontend/src/routes/workspaced_content.rs
@@ -1,66 +1,19 @@
+use leptos_query::QueryResult;
 use leptos_router::{Outlet, ProtectedRoute};
 
-use crate::{pages::*, prelude::*, utils::AuthState};
+use crate::{
+	pages::*,
+	prelude::*,
+	queries::{list_workspaces_query, AllWorkspacesTag},
+	utils::AuthState,
+};
 
 /// The view for the Workspaced Routes
 #[component]
 pub fn WorkspacedRouteView() -> impl IntoView {
-	let (state, _) = AuthState::load();
-	let current_workspace_id = Signal::derive(move || state.get().get_last_used_workspace_id());
-	let (_, set_current_workspace) = create_signal(None);
-
-	// Don't think this will work properly. Need to check
-	create_effect(move |_| {
-		set_current_workspace.set(current_workspace_id.get());
-	});
-
-	let workspace_list = create_resource(
-		move || state.get().get_access_token(),
-		move |value| async move { list_user_workspace(value).await },
-	);
-
-	let current_workspace = Signal::derive(move || {
-		if let Some(workspace_id) = current_workspace_id.get() {
-			workspace_list
-				.get()
-				.map(|list| {
-					list.ok().map(|list| {
-						list.workspaces
-							.iter()
-							.find(|&x| x.id == workspace_id)
-							.cloned()
-					})
-				})
-				.flatten()
-				.flatten()
-		} else {
-			None
-		}
-	});
-
 	view! {
 		<Sidebar>
-			<Transition>
-				{
-					move || match workspace_list.get() {
-						Some(workspace_list) => {
-							match workspace_list {
-								Ok(data) => {
-									view! {
-										<WorkspaceCard
-											current_workspace={current_workspace}
-											set_workspace_id={set_current_workspace}
-											workspaces={data.clone().workspaces}
-										/>
-									}.into_view()
-								},
-								Err(_) => view! {"Error Loading"}.into_view()
-							}
-						},
-						None => view! {"loading..."}.into_view()
-					}
-				}
-			</Transition>
+			<WorkspaceSidebarComponent/>
 		</Sidebar>
 
 		<main class="fc-fs-ct full-width px-lg">
@@ -74,10 +27,10 @@ pub fn WorkspacedRouteView() -> impl IntoView {
 pub fn WorkspacedRoutes() -> impl IntoView {
 	let (state, set_state) = AuthState::load();
 
-	let workspace_list = create_resource(
-		move || state.get().get_access_token(),
-		move |value| async move { list_user_workspace(value).await },
-	);
+	let QueryResult {
+		data: workspace_list,
+		..
+	} = list_workspaces_query().use_query(|| AllWorkspacesTag);
 
 	let current_workspace_id =
 		Signal::derive(move || match state.get().get_last_used_workspace_id() {
