+++ runners/common/src/routes/workspace/deployment/list_deployment.rs
@@ -54,10 +54,9 @@ pub async fn list_deployment(
 		.map(|row| {
 			let deployment_id = row.try_get::<Uuid, _>("id")?;
 			let name = row.try_get::<String, _>("name")?;
-			let image_tag = row.try_get::<String, _>("image_tag")?;
 			let status = row.try_get::<DeploymentStatus, _>("status")?;
 			let registry = row.try_get::<String, _>("registry")?;
-			let repository_id = row.try_get::<Uuid, _>("repository_id")?;
+			let image_tag = row.try_get::<String, _>("image_tag")?;
 			let image_name = row.try_get::<String, _>("image_name")?;
 			let machine_type = row.try_get::<Uuid, _>("machine_type")?;
 
@@ -70,7 +69,7 @@ pub async fn list_deployment(
 					registry: if registry == PatrRegistry.to_string() {
 						DeploymentRegistry::PatrRegistry {
 							registry: PatrRegistry,
-							repository_id,
+							repository_id: Uuid::nil(),
 						}
 					} else {
 						DeploymentRegistry::ExternalRegistry {
